services/api/requirements.txt
Flask==2.1.2
flask-restx==0.5.1
Flask-Cors==3.0.10
gunicorn==20.1.0
Werkzeug==2.1.2
click==8.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pytz==2022.1
six==1.16.0
python-dotenv==0.19.2
python-decouple==3.5
openai